video caption parsing attempts
hopefully this works. should regex each unit's combined captions. if it finds something, then it will regex each video within the unit for the query and return the video it is in. if it finds multiple things it will return all the videos in the unit. if it finds nothing it will try to return the unit's videos based on wikisearch timperiod or sum